backend/requirement.txt
langchain==0.3.0
langchain-community==0.3.0
langchain-core==0.3.0

# Google Gemini + LangChain
langchain-google-genai==1.0.7
google-generativeai==0.7.2
google-ai-generativelanguage==0.6.15   # works with 0.7.x

# Database
mysql-connector-python==9.0.0

# Env
python-dotenv==1.0.1

# Vector DB
faiss-cpu==1.8.0
